refactor input
