+++ core/amber/src/main/scala/edu/uci/ics/amber/engine/architecture/messaginglayer/InputManager.scala
@@ -1,17 +1,15 @@
 package edu.uci.ics.amber.engine.architecture.messaginglayer
 
 import edu.uci.ics.amber.engine.common.AmberLogging
-import edu.uci.ics.amber.engine.common.virtualidentity.{ActorVirtualIdentity, ChannelIdentity}
+import edu.uci.ics.amber.engine.common.virtualidentity.ActorVirtualIdentity
 import edu.uci.ics.amber.engine.common.workflow.PortIdentity
-import edu.uci.ics.texera.workflow.common.tuple.Tuple
 import edu.uci.ics.texera.workflow.common.tuple.schema.Schema
 
 import scala.collection.mutable
 
 class InputManager(val actorId: ActorVirtualIdentity) extends AmberLogging {
-  private var inputBatch: Array[Tuple] = _
-  private var currentInputIdx: Int = -1
-  var currentChannelId: ChannelIdentity = _
+
+
 
   private val ports: mutable.HashMap[PortIdentity, WorkerPort] = mutable.HashMap()
   def getAllPorts: Set[PortIdentity] = {
@@ -36,25 +34,5 @@ class InputManager(val actorId: ActorVirtualIdentity) extends AmberLogging {
     this.ports(portId).channels.values.forall(completed => completed)
   }
 
-  def hasUnfinishedInput: Boolean = inputBatch != null && currentInputIdx + 1 < inputBatch.length
 
-  def getNextTuple: Tuple = {
-    currentInputIdx += 1
-    inputBatch(currentInputIdx)
-  }
-  def getCurrentTuple: Tuple = {
-    if (inputBatch == null) {
-      null
-    } else if (inputBatch.isEmpty) {
-      null // TODO: create input exhausted
-    } else {
-      inputBatch(currentInputIdx)
-    }
-  }
-
-  def initBatch(channelId: ChannelIdentity, batch: Array[Tuple]): Unit = {
-    currentChannelId = channelId
-    inputBatch = batch
-    currentInputIdx = -1
-  }
 }
