+++ core/amber/src/main/scala/edu/uci/ics/amber/engine/architecture/worker/TupleProcessingManager.scala
@@ -3,8 +3,9 @@ package edu.uci.ics.amber.engine.architecture.worker
 import edu.uci.ics.amber.engine.architecture.worker.DataProcessor.{FinalizeExecutor, FinalizePort}
 import edu.uci.ics.amber.engine.architecture.worker.TupleProcessingManager.DPOutputIterator
 import edu.uci.ics.amber.engine.common.tuple.amber.TupleLike
-import edu.uci.ics.amber.engine.common.virtualidentity.ActorVirtualIdentity
+import edu.uci.ics.amber.engine.common.virtualidentity.{ActorVirtualIdentity, ChannelIdentity}
 import edu.uci.ics.amber.engine.common.workflow.PortIdentity
+import edu.uci.ics.texera.workflow.common.tuple.Tuple
 
 import scala.collection.mutable
 
@@ -37,7 +38,32 @@ object TupleProcessingManager{
   }
 }
 class TupleProcessingManager( val actorId: ActorVirtualIdentity) {
+  private var inputBatch: Array[Tuple] = _
+  private var currentInputIdx: Int = -1
+  var currentChannelId: ChannelIdentity = _
   val outputIterator: DPOutputIterator = new DPOutputIterator()
+  def hasUnfinishedInput: Boolean = inputBatch != null && currentInputIdx + 1 < inputBatch.length
+
+  def getNextTuple: Tuple = {
+    currentInputIdx += 1
+    inputBatch(currentInputIdx)
+  }
+  def getCurrentTuple: Tuple = {
+    if (inputBatch == null) {
+      null
+    } else if (inputBatch.isEmpty) {
+      null // TODO: create input exhausted
+    } else {
+      inputBatch(currentInputIdx)
+    }
+  }
+
+  def initBatch(channelId: ChannelIdentity, batch: Array[Tuple]): Unit = {
+    currentChannelId = channelId
+    inputBatch = batch
+    currentInputIdx = -1
+  }
+
 
   def hasUnfinishedOutput: Boolean = outputIterator.hasNext
