+++ core/amber/src/main/scala/edu/uci/ics/amber/engine/architecture/worker/promisehandlers/QueryCurrentInputTupleHandler.scala
@@ -13,6 +13,6 @@ trait QueryCurrentInputTupleHandler {
   this: DataProcessorRPCHandlerInitializer =>
 
   registerHandler { (msg: QueryCurrentInputTuple, sender) =>
-    dp.inputManager.getCurrentTuple
+    dp.tupleProcessingManager.getCurrentTuple
   }
 }
